refactor(aspida): clarify cli command
