--- packages/aspida/src/cli/command.ts
@@ -1,9 +1,0 @@
-export type Command = {
-  exec(): void
-}
-
-export const nullCommand: Command = {
-  exec() {
-    // nothing to do
-  }
-}
